// src/components/Txt/Txt.enums.ts
export enum TxtType {
    SMALL = 'small',
    MEDIUM = 'medium',
    BIG = 'big',
}

export enum TxtTag {
    SPAN = 'span',
    H1 = 'h1',
    H2 = 'h2',
    H3 = 'h3',
    H4 = 'h4',
    H5 = 'h5',
    H6 = 'h6',
    PARAGRAPH = 'p',
}

export enum TxtColor {
    WHITE = 'white',
}

export enum TxtAlign {
    LEFT = 'left',
    CENTER = 'center',
    RIGHT = 'right',
}

export enum TxtWeight {
    LIGHT = 'text-weight_light',
    MEDIUM = 'text-weight_medium',
    BOLD = 'text-weight_bold',
}